Move email reply logic entirely to lib/api
Delete temporal/agent_loop/activities/email_reply.ts since those
functions are not Temporal activities - the real activities are in
finalize.ts which now imports directly from lib/api/assistant/email_reply.ts.
