+++ front/lib/api/assistant/email_reply.ts
@@ -11,12 +11,12 @@ import {
   replyToEmail,
 } from "@app/lib/api/assistant/email_trigger";
 import config from "@app/lib/api/config";
-import type { Authenticator } from "@app/lib/auth";
+import type { AuthenticatorType } from "@app/lib/auth";
 import { getConversationRoute } from "@app/lib/utils/router";
 import logger from "@app/logger/logger";
 import { isDevelopment } from "@app/types";
-
-const PRODUCTION_DUST_WORKSPACE_ID = config.getProductionDustWorkspaceId();
+import type { AgentLoopArgs } from "@app/types/assistant/agent_run";
+import { getAgentLoopData } from "@app/types/assistant/agent_run";
 
 /**
  * Reconstructs a minimal InboundEmail from the stored context.
@@ -57,10 +57,11 @@ function checkEmailReplyGating(
     );
     return false;
   }
+  const productionDustWorkspaceId = config.getProductionDustWorkspaceId();
   if (
     !isDevelopment() &&
-    PRODUCTION_DUST_WORKSPACE_ID &&
-    context.workspaceId !== PRODUCTION_DUST_WORKSPACE_ID
+    productionDustWorkspaceId &&
+    context.workspaceId !== productionDustWorkspaceId
   ) {
     logger.warn(
       { agentMessageId, workspaceId: context.workspaceId },
@@ -75,36 +76,47 @@ function checkEmailReplyGating(
  * Send an email reply after agent message completion.
  * Fire-and-forget: failures are logged but don't throw.
  */
-export async function sendEmailReplyOnCompletion({
-  auth,
-  workspaceId,
-  agentMessageId,
-  agentMessageContent,
-  conversationId,
-}: {
-  auth: Authenticator;
-  workspaceId: string;
-  agentMessageId: string;
-  agentMessageContent: string | null;
-  conversationId: string;
-}): Promise<void> {
+export async function sendEmailReplyOnCompletion(
+  authType: AuthenticatorType,
+  agentLoopArgs: AgentLoopArgs
+): Promise<void> {
   try {
+    // Only process email-originated messages.
+    if (agentLoopArgs.userMessageOrigin !== "email") {
+      return;
+    }
+
     const context = await getAndDeleteEmailReplyContext(
-      workspaceId,
-      agentMessageId
+      authType.workspaceId,
+      agentLoopArgs.agentMessageId
     );
     if (!context) {
       // Context not found - either expired, never stored, or already processed.
       logger.info(
-        { agentMessageId },
+        { agentMessageId: agentLoopArgs.agentMessageId },
         "[email] No email reply context found, skipping reply"
       );
       return;
     }
 
-    if (!checkEmailReplyGating(context, agentMessageId)) {
+    if (!checkEmailReplyGating(context, agentLoopArgs.agentMessageId)) {
       return;
     }
+
+    // Get the completed agent message data.
+    const dataRes = await getAgentLoopData(authType, agentLoopArgs);
+    if (dataRes.isErr()) {
+      logger.warn(
+        {
+          agentMessageId: agentLoopArgs.agentMessageId,
+          error: dataRes.error,
+        },
+        "[email] Failed to get agent loop data for email reply"
+      );
+      return;
+    }
+
+    const { auth, agentMessage, conversation } = dataRes.value;
 
     // Get agent configuration for the reply sender name.
     const agentConfiguration = await getAgentConfiguration(auth, {
@@ -114,7 +126,7 @@ export async function sendEmailReplyOnCompletion({
 
     // Render the agent message content as HTML.
     const htmlContent = sanitizeHtml(
-      await marked.parse(agentMessageContent ?? ""),
+      await marked.parse(agentMessage.content ?? ""),
       {
         allowedTags: sanitizeHtml.defaults.allowedTags.concat(["img"]),
       }
@@ -123,7 +135,7 @@ export async function sendEmailReplyOnCompletion({
     // Build the full HTML with conversation link.
     const conversationLink = getConversationRoute(
       context.workspaceId,
-      conversationId,
+      conversation.sId,
       undefined,
       config.getClientFacingUrl()
     );
@@ -139,8 +151,8 @@ export async function sendEmailReplyOnCompletion({
 
     logger.info(
       {
-        agentMessageId,
-        conversationId,
+        agentMessageId: agentLoopArgs.agentMessageId,
+        conversationId: conversation.sId,
         to: context.fromEmail,
       },
       "[email] Sent email reply on agent completion"
@@ -149,7 +161,7 @@ export async function sendEmailReplyOnCompletion({
     logger.warn(
       {
         err,
-        agentMessageId,
+        agentMessageId: agentLoopArgs.agentMessageId,
       },
       "[email] Failed to send email reply on completion, skipping"
     );
@@ -160,29 +172,29 @@ export async function sendEmailReplyOnCompletion({
  * Send an error email on agent error/cancellation.
  * Fire-and-forget: failures are logged but don't throw.
  */
-export async function sendEmailReplyOnError({
-  workspaceId,
-  agentMessageId,
-  errorMessage,
-}: {
-  workspaceId: string;
-  agentMessageId: string;
-  errorMessage: string;
-}): Promise<void> {
+export async function sendEmailReplyOnError(
+  authType: AuthenticatorType,
+  agentLoopArgs: AgentLoopArgs,
+  errorMessage: string
+): Promise<void> {
   try {
+    if (agentLoopArgs.userMessageOrigin !== "email") {
+      return;
+    }
+
     const context = await getAndDeleteEmailReplyContext(
-      workspaceId,
-      agentMessageId
+      authType.workspaceId,
+      agentLoopArgs.agentMessageId
     );
     if (!context) {
       logger.info(
-        { agentMessageId },
+        { agentMessageId: agentLoopArgs.agentMessageId },
         "[email] No email reply context found for error reply, skipping"
       );
       return;
     }
 
-    if (!checkEmailReplyGating(context, agentMessageId)) {
+    if (!checkEmailReplyGating(context, agentLoopArgs.agentMessageId)) {
       return;
     }
 
@@ -197,7 +209,7 @@ export async function sendEmailReplyOnError({
 
     logger.info(
       {
-        agentMessageId,
+        agentMessageId: agentLoopArgs.agentMessageId,
         to: context.fromEmail,
         errorMessage,
       },
@@ -207,7 +219,7 @@ export async function sendEmailReplyOnError({
     logger.warn(
       {
         err,
-        agentMessageId,
+        agentMessageId: agentLoopArgs.agentMessageId,
       },
       "[email] Failed to send error email reply, skipping"
     );
